requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==4.2.1
django-allauth==0.52.0
django-filter==23.2
django-model-utils==4.3.1
djangorestframework==3.14.0
docopt==0.6.2
idna==3.4
install==1.3.5
oauthlib==3.2.2
pipreqs==0.4.13
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
urllib3==2.0.2
yarg==0.1.9
